+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/batched_transition/document_create_transition/validate_structure/mod.rs
@@ -7,7 +7,7 @@ use platform_version::version::PlatformVersion;
 
 mod v0;
 
-pub trait DocumentCreateTransitionStructureValidation {
+pub(crate) trait DocumentCreateTransitionStructureValidation {
     fn validate_structure(
         &self,
         owner_id: Identifier,