+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/batched_transition/document_delete_transition/validate_structure/mod.rs
@@ -7,7 +7,7 @@ use platform_version::version::PlatformVersion;
 
 mod v0;
 
-pub trait DocumentDeleteTransitionStructureValidation {
+pub(crate) trait DocumentDeleteTransitionStructureValidation {
     fn validate_structure(
         &self,
         document_type: DocumentTypeRef,